abstract constants into files
