+++ src/apps/events/constants.js
@@ -1,0 +1,7 @@
+const DEFAULT_COLLECTION_QUERY = {
+  sortby: 'modified_on:desc',
+}
+
+module.exports = {
+  DEFAULT_COLLECTION_QUERY,
+}
