+++ src/apps/events/router.js
@@ -1,15 +1,15 @@
 const router = require('express').Router()
 
 const { setDefaultQuery } = require('../middleware')
+const { GLOBAL_NAV_ITEMS } = require('../constants')
+const { DEFAULT_COLLECTION_QUERY } = require('./constants')
+
 const { renderDetailsPage } = require('./controllers/details')
 const { renderEditPage } = require('./controllers/edit')
 const { postDetails, getEventDetails } = require('./middleware/details')
 const { getRequestBody, getEventsCollection } = require('./middleware/collection')
 const { renderEventList } = require('./controllers/list')
 
-const DEFAULT_COLLECTION_QUERY = {
-  sortby: 'modified_on:desc',
-}
 
 router.param('id', getEventDetails)
