+++ src/apps/interactions/constants.js
@@ -1,0 +1,7 @@
+const DEFAULT_COLLECTION_QUERY = {
+  sortby: 'date:desc',
+}
+
+module.exports = {
+  DEFAULT_COLLECTION_QUERY,
+}
