+++ src/apps/investment-projects/router.js
@@ -1,6 +1,9 @@
 const router = require('express').Router()
 
 const { setLocalNav, setDefaultQuery, redirectToFirstNavItem } = require('../middleware')
+const { GLOBAL_NAV_ITEMS } = require('../constants')
+const { DEFAULT_COLLECTION_QUERY, LOCAL_NAV } = require('./constants')
+
 const { shared } = require('./middleware')
 const {
   getBriefInvestmentSummary,
@@ -56,23 +59,6 @@ const {
 
 const interactionsRouter = require('../interactions/router.sub-app')
 
-const LOCAL_NAV = [
-  { path: 'details', label: 'Project details' },
-  { path: 'team', label: 'Project team' },
-  { path: 'interactions', label: 'Interactions' },
-  { path: 'evaluation', label: 'Evaluations' },
-  { path: 'audit', label: 'Audit history' },
-  { path: 'documents', label: 'Documents' },
-]
-
-const currentYear = (new Date()).getFullYear()
-const DEFAULT_COLLECTION_QUERY = {
-  estimated_land_date_after: `${currentYear}-04-05`,
-  estimated_land_date_before: `${currentYear + 1}-04-06`,
-  sortby: 'estimated_land_date:asc',
-}
-
-router.use('/:investmentId', setLocalNav(LOCAL_NAV))
 
 router.param('investmentId', shared.getInvestmentDetails)
 router.param('companyId', shared.getCompanyDetails)
